Split data processing and formatting for time table generation.
This commit introduces the opening_hours_html_formatter.js library
which converts opening hours (intervals) into HTML elements (Week).

- OpeningTimes
- WeekGenerator

The generation of HTML being displayed to the user is a second processing
step which makes use of the following class:

- WeekTableHtmlGenerator

# Motivation

The motivation for this refactoring is to finally extract the conversion
and formatting of data retrieved from the opening_hours.js library into a
separate library to allow re-use in other projects.